WIP: MessagePane refactoring and scaling
